imagen libro default y borrado correcto

+++ main/models.py
@@ -1,3 +1,4 @@
+from email.policy import default
 from tabnanny import verbose
 from django.db import models
 
@@ -19,22 +20,24 @@ class Categoria(models.Model):
 class Historia(models.Model):
     id = models.AutoField(primary_key=True)
     titulo = models.CharField(max_length=100, verbose_name='titulo')
-    portada = models.ImageField(upload_to='imagenes/portadas/', verbose_name='portada', null=True)
+    portada = models.ImageField(upload_to='imagenes/portadas/', default="imagenes/portadas/book-default.png", verbose_name='portada')
     descripcion = models.CharField(max_length=100, verbose_name='descripcion', null=True)
+    id_categoria = models.ManyToManyField(Categoria)
+
 
     def delete(self, using=None, keep_parents=False):
-        self.portada.storage.delete(self.portada.name)
+        if self.portada.name != 'imagenes/portadas/book-default.png':
+            self.portada.storage.delete(self.portada.name)
         super().delete()
     
     def get_portada(self):
         return self.portada.name
 
     def del_portada(self, borrar):
-        self.portada.storage.delete(borrar)
+        if borrar != 'imagenes/portadas/book-default.png':
+            self.portada.storage.delete(borrar)
+
+
+    def __str__(self) -> str:
+        return f'{self.id} {self.titulo}'
         
-
-
-class CategoriaDeHistoria(models.Model):
-    id = models.AutoField(primary_key=True)
-    id_historia = models.IntegerField(verbose_name='id_historia')
-    id_categoria = models.IntegerField(verbose_name='id_categoria')
